correct int64_t in kMeans.h

+++ headers/kMeans.h
@@ -6,7 +6,7 @@
 typedef struct {
     cluster_t *clusters;
     point_t **points;
-    u_int32_t dimension;
+    int64_t dimension;
     int64_t size;
 } k_means_t;
